# Raise an informative error message when object array has mixed types