--- a/xarray/conventions.py
+++ b/xarray/conventions.py
@@ -67,7 +67,7 @@ def _infer_dtype(array, name: T_Name = None) -> np.dtype:
         raise ValueError(
             "unable to infer dtype on variable {!r}; object array "
             "contains mixed native types: {}".format(
-                name, ",".join(map(lambda x: x.__name__, native_dtypes))
+                name, ", ".join(map(lambda x: x.__name__, native_dtypes))
             )
         )