--- a/xarray/tests/test_conventions.py
+++ b/xarray/tests/test_conventions.py
@@ -495,6 +495,12 @@ class TestCFEncodedDataStore(CFEncodedBase):
         pass
 
 
+def test_infer_dtype_error_on_mixed_types():
+    data = np.array([["x", 1], ["y", 2]], dtype="object")
+    with pytest.raises(ValueError):
+        conventions._infer_dtype(data, "test")
+
+
 class TestDecodeCFVariableWithArrayUnits:
     def test_decode_cf_variable_with_array_units(self) -> None:
         v = Variable(["t"], [1, 2, 3], {"units": np.array(["foobar"], dtype=object)})