--- a/xarray/tests/test_conventions.py
+++ b/xarray/tests/test_conventions.py
@@ -495,6 +495,12 @@ class TestCFEncodedDataStore(CFEncodedBase):
         pass
 
 
+def test_infer_dtype_error_on_mixed_types():
+    data = np.array([["x", 1], ["y", 2]], dtype="object")
+    with pytest.raises(ValueError):
+        conventions._infer_dtype(data, "test")
+
+
 @pytest.mark.parametrize(
     "data",
     [