@pep8speaks 2020-12-17
Hello @andersy005! Thanks for updating this PR. We checked the lines you've touched for [PEP 8](https://www.python.org/dev/peps/pep-0008) issues, and found:

* In the file [`xarray/conventions.py`](https://github.com/pydata/xarray/blob/b3ecf0c3df817def11f04889478783e661aa79ca/xarray/conventions.py):

> [Line 142:61](https://github.com/pydata/xarray/blob/b3ecf0c3df817def11f04889478783e661aa79ca/xarray/conventions.py#L142): [F821](https://duckduckgo.com/?q=pep8%20F821) undefined name 'a'