@andersy005 2020-12-17
> Alternatives — not ideal ones — would be to wait until the main error is raised, or only test a subset of the values. 

I thought of taking a random sample from the array and checking the types on the sample only, but I wasn't so confident about how representative this sample would be and/or how to deal with misleading, skewed samples. If anyone has thoughts on this, please let me know.