@mathause 2020-12-17
It took around 40 s for an array of 10**9 elements. That would be around 150 years of daily data (180\*360\*150\*365). I am not sure though how much sense it makes to have such a large array with object dtype. Also an array of this size is likely a dask array and there is already a performance warning on this. So I'd say go ahead. 

https://github.com/pydata/xarray/blob/68d3c340304ccc57cac569b22781e2b9c3dac913/xarray/conventions.py#L194-L197