@mathause 2021-01-21
Yes, I'd say go ahead. (I just hope it's not too big of a performance hit for normal use cases.)